Get it all somewhat working
