+++ gulp/config.js
@@ -1,0 +1,7 @@
+var path = require('path');
+
+module.exports = {
+  static_dir: path.resolve(__dirname, '../..'),
+  polymer_dir: path.resolve(__dirname, '..'),
+  build_dir: path.resolve(__dirname, '../build'),
+};
