+++ gulp/tasks/clean.js
@@ -1,0 +1,6 @@
+import del from 'del';
+import gulp from 'gulp';
+
+gulp.task('clean', () => {
+  return del(['build', 'build-temp']);
+});
