+++ gulp/tasks/default.js
@@ -1,0 +1,9 @@
+import gulp from 'gulp';
+import runSequence from 'run-sequence';
+
+gulp.task('default', () => {
+  return runSequence.use(gulp)(
+    'clean',
+    'build',
+  );
+});
