+++ gulp/tasks/service-worker.js.tmpl
