+++ package.json
@@ -9,15 +9,10 @@
   "scripts": {
     "clean": "rm -rf build/* build-temp/*",
     "gulp": "gulp",
-    "js_dev": "node script/gen-service-worker.js && npm run gulp watch_ru_all",
-    "js_dev_demo": "BUILD_DEMO=1 npm run gulp watch_ru_all",
-    "js_prod": "BUILD_DEV=0 npm run gulp ru_all",
-    "js_demo": "BUILD_DEV=0 BUILD_DEMO=1 npm run gulp ru_all",
-    "frontend_html": "node script/vulcanize.js",
-    "frontend_prod": "npm run js_prod && npm run frontend_html",
-    "frontend_demo": "npm run js_demo && npm run frontend_html",
-    "ru_all": "npm run gulp ru_all",
-    "watch_ru_all": "npm run gulp watch_ru_all",
+    "build": "BUILD_DEV=0 gulp",
+    "build_demo": "BUILD_DEV=0 BUILD_DEMO=1 gulp",
+    "dev": "npm run gulp ru_all",
+    "dev-watch": "npm run gulp watch_ru_all",
     "lint_js": "eslint src panels preact-src --ext js,html",
     "lint_html": "ls -1 src/home-assistant.html panels/**/ha-panel-*.html | xargs polymer lint --input",
     "test": "npm run lint_js && npm run lint_html"
@@ -56,6 +51,7 @@
     "polymer-bundler": "^3.0.1",
     "polymer-cli": "^1.3.1",
     "polymer-lint": "^0.8.3",
+    "require-dir": "^0.3.2",
     "rollup": "^0.45.2",
     "rollup-plugin-babel": "^2.7.1",
     "rollup-plugin-commonjs": "^8.0.2",
@@ -63,6 +59,7 @@
     "rollup-plugin-replace": "^1.1.1",
     "rollup-plugin-uglify": "^2.0.1",
     "rollup-watch": "^4.3.1",
+    "run-sequence": "^2.1.0",
     "sw-precache": "^5.2.0",
     "uglify-js": "^3.0.26",
     "vulcanize": "^1.16.0",
